Bug fixes
Limited the portion of ground velocity that a body applies upon impact with a mover. This makes sure that the body stays "stuck" to the mover upon landing instead of being given excess upwards motion.

Also added a condition that skips the mover collision detection if the body's motion relative to the mover would result in the body "outpacing" the mover within the next frame. This condition is necessary because movers always move before bodies in the simulation frame, so even if the body is moving away from the mover, it may still register a collision.